fix: handlers not being registered properly
