+++ src/main/handlers.ts
@@ -1,10 +1,13 @@
-import { Car, PartialValue, Stint, Tire, Track } from '../shared/model';
+import { Car, ImportData, PartialValue, Stint, Tire, Track } from '../shared/model';
 import { randomUUID } from 'crypto';
 import {
   deleteCarId,
   deleteStintId,
   deleteTireId,
   deleteTrackId,
+  getAllTableNames,
+  getQueryAllTables,
+  importDataWithOptions,
   insertCar,
   insertStint,
   insertTire,
@@ -22,6 +25,8 @@ import { tireSchema } from '../shared/schema/tireSchema';
 import { carSchema } from '../shared/schema/carSchema';
 import { trackSchema } from '../shared/schema/trackSchema';
 import { stintSchema } from '../shared/schema/stintSchema';
+import { app, BrowserWindow, dialog } from 'electron';
+import fs from 'fs';
 
 const getTracks = (): Track[] => {
   return queryTracks.all();
@@ -171,6 +176,72 @@ const deleteStint = (_, stintId: string) => {
   deleteStintId.run(stintId);
 };
 
+async function selectImportFile() {
+  const focusedWindow = BrowserWindow.getFocusedWindow();
+  if (!focusedWindow) return { canceled: true };
+
+  const result = await dialog.showOpenDialog(focusedWindow, {
+    title: 'Import Database',
+    filters: [{ name: 'JSON Files', extensions: ['json'] }],
+    properties: ['openFile']
+  });
+  if (result.canceled || result.filePaths.length === 0) return { canceled: true };
+  const filePath = result.filePaths[0];
+  const fileContent = fs.readFileSync(filePath, 'utf-8');
+  return { canceled: false, fileContent };
+}
+
+function confirmImport(_, importedData: ImportData) {
+  // Validate and import as before
+  importDataWithOptions(importedData, { conflictResolution: 'merge' });
+  return { success: true };
+}
+
+async function exportData() {
+  try {
+    const focusedWindow = BrowserWindow.getFocusedWindow();
+    if (!focusedWindow) return;
+
+    // Show save dialog to let user choose location and filename
+    const result = await dialog.showSaveDialog(focusedWindow, {
+      title: 'Export Database',
+      defaultPath: `tire-logger-export-${new Date().toISOString().split('T')[0]}.json`,
+      filters: [
+        { name: 'JSON Files', extensions: ['json'] },
+        { name: 'All Files', extensions: ['*'] }
+      ]
+    });
+
+    // User cancelled the dialog
+    if (result.canceled) {
+      return { success: false, message: undefined };
+    }
+
+    // Get the selected file path
+    const filePath = result.filePath;
+
+    // Export data logic
+    const tables = getAllTableNames.all();
+    const data = getQueryAllTables([...tables.map((table) => table.name)]);
+
+    // Add metadata to export
+    const exportData = {
+      exportDate: new Date().toISOString(),
+      version: app.getVersion(),
+      appName: 'tire-logger',
+      data: data
+    };
+
+    // Write to the user-selected file
+    fs.writeFileSync(filePath, JSON.stringify(exportData, null, 2));
+    console.log(`Data exported successfully to ${filePath}`);
+
+    return { success: true, filePath };
+  } catch (error) {
+    return { success: false, message: error };
+  }
+}
+
 export const handlers = [
   // Tracks
   getTracks,
@@ -187,5 +258,9 @@ export const handlers = [
   // Stints
   putStint,
   getStints,
-  deleteStint
+  deleteStint,
+  // Data Transfer
+  selectImportFile,
+  confirmImport,
+  exportData
 ];
